backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
spotipy>=2.23.0
python-dotenv>=1.0.0
lyricsgenius>=3.0.0
transformers>=4.30.0
torch>=2.0.0
requests>=2.31.0
librosa>=0.10.0
soundfile>=0.12.0
spotify-preview-finder>=1.0.0
librosa>=0.10.0
soundfile>=0.12.0
numba>=0.57.0

